use ints rather than longs
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@18814 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ uvm-lib/api/com/untangle/uvm/message/RemoteMessageManager.java
@@ -42,8 +42,8 @@ import com.untangle.uvm.security.Tid;
 public interface RemoteMessageManager
 {
     MessageQueue getMessageQueue();
-    MessageQueue getMessageQueue(Long key);
-    MessageQueue getMessageQueue(Long key, Policy p);
+    MessageQueue getMessageQueue(Integer key);
+    MessageQueue getMessageQueue(Integer key, Policy p);
     StatDescs getStatDescs(Tid t);
     Stats getStats(Tid t);
     Map<String, Object> getSystemStats();
@@ -51,5 +51,5 @@ public interface RemoteMessageManager
     void setActiveMetrics(Tid tid, List<ActiveStat> activeMetrics);
     List<Message> getMessages();
 
-    Long getMessageKey();
+    Integer getMessageKey();
 }
